requirements.txt
airsim
opencv-python
numpy
pandas
matplotlib
tqdm
pyyaml
torch
gymnasium
stable-baselines3
tensorboard
shimmy>=0.2.1